backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
openai>=1.0
aiosqlite>=0.20
httpx>=0.27
feedparser>=6.0
apscheduler>=3.10
python-dotenv>=1.0
hyperliquid-python-sdk>=0.9.0
eth-account>=0.10.0
